chore(vue): reorganize Menu tests
