+++ packages/vue/src/components/Menu/tests/assertions.ts
